+++ src/BuildBot.CloudFormation/Services/AwsCloudFormation.cs
@@ -1,13 +1,9 @@
 using System;
-using System.Diagnostics.CodeAnalysis;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
-using Amazon;
 using Amazon.CloudFormation;
 using Amazon.CloudFormation.Model;
-using Amazon.Runtime;
-using BuildBot.CloudFormation.Configuration;
 using BuildBot.CloudFormation.Services.LoggingExtensions;
 using Microsoft.Extensions.Logging;
 
@@ -15,23 +11,13 @@ namespace BuildBot.CloudFormation.Services;
 
 public sealed class AwsCloudFormation : IAwsCloudFormation
 {
-    private readonly BasicAWSCredentials _credentials;
-    private readonly RegionEndpoint _endpoint;
+    private readonly IAmazonCloudFormation _cloudFormationClient;
     private readonly ILogger<AwsCloudFormation> _logger;
-    private readonly SnsNotificationOptions _options;
 
-    [SuppressMessage(
-        category: "Roslynator.Analyzers",
-        checkId: "RCS1231: Make parameter ref read-only.",
-        Justification = "DI"
-    )]
-    public AwsCloudFormation(SnsNotificationOptions options, ILogger<AwsCloudFormation> logger)
+    public AwsCloudFormation(IAmazonCloudFormation cloudFormationClient, ILogger<AwsCloudFormation> logger)
     {
-        this._options = options;
+        this._cloudFormationClient = cloudFormationClient;
         this._logger = logger;
-
-        this._endpoint = RegionEndpoint.GetBySystemName(this._options.Region);
-        this._credentials = new(accessKey: this._options.AccessKey, secretKey: this._options.SecretKey);
     }
 
     public async ValueTask<StackDetails?> GetStackDetailsAsync(
@@ -41,45 +27,32 @@ public sealed class AwsCloudFormation : IAwsCloudFormation
     {
         try
         {
-            using (
-                AmazonCloudFormationClient cloudFormationClient = new(
-                    credentials: this._credentials,
-                    region: this._endpoint
-                )
-            )
+            DescribeStacksRequest request = new() { StackName = deployment.StackName };
+
+            DescribeStacksResponse result = await this._cloudFormationClient.DescribeStacksAsync(
+                request: request,
+                cancellationToken: cancellationToken
+            );
+
+            Stack? stack = result.Stacks.FirstOrDefault();
+
+            if (stack is null)
             {
-                DescribeStacksRequest request = new() { StackName = deployment.StackName };
-
-                DescribeStacksResponse? result = await cloudFormationClient.DescribeStacksAsync(
-                    request: request,
-                    cancellationToken: cancellationToken
-                );
-
-                if (result is null)
-                {
-                    return null;
-                }
-
-                Stack? stack = result.Stacks.FirstOrDefault();
-
-                if (stack is null)
-                {
-                    return null;
-                }
-
-                string projectDescription = stack.Description;
-
-                Tag? versionTag = stack.Tags.Find(t => StringComparer.Ordinal.Equals(x: t.Key, y: "Version"));
-
-                string? projectVersion = versionTag?.Value;
-
-                if (!string.IsNullOrWhiteSpace(projectDescription) || !string.IsNullOrWhiteSpace(projectVersion))
-                {
-                    return new StackDetails(Description: projectDescription, Version: projectVersion);
-                }
-
                 return null;
             }
+
+            string projectDescription = stack.Description;
+
+            Tag? versionTag = stack.Tags.Find(t => StringComparer.Ordinal.Equals(x: t.Key, y: "Version"));
+
+            string? projectVersion = versionTag?.Value;
+
+            if (!string.IsNullOrWhiteSpace(projectDescription) || !string.IsNullOrWhiteSpace(projectVersion))
+            {
+                return new StackDetails(Description: projectDescription, Version: projectVersion);
+            }
+
+            return null;
         }
         catch (Exception exception)
         {